Move isRecyclable into scratchblocksutils
